+++ src/utilities.ts
@@ -64,6 +64,19 @@ export function xmlArray<T>(xml: any, ...path: string[]): T[] {
 const ok = Object.keys
 export const xmlRoot = (o: any) => o[ok(o)[0]]
 
+export const stripNs = (x: any) =>
+  x &&
+  ok(x).reduce((obj, key) => {
+    const nk =
+      key
+        .split(":")
+        .slice(1)
+        .join(":") || key
+    if (nk in obj) obj[key] = key
+    else obj[nk] = x[key]
+    return obj
+  }, {} as any)
+
 // extract XML attributes of a node from its JSON representation
 export const xmlNodeAttr = (n: any) =>
   n &&