+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## Unreleased
+
+### Added
+
+- AbapGit stage
+- AbapGit push
+
+### Changed
+
+- adopted semantyc versioning
+
 ## [0.5.37] 2020-03-23
 
 ### Added